--- camera-test.py
@@ -1,81 +1,0 @@
-#!/usr/bin/python3
-
-# ------------------------------------------------------------------------------
-# Display the image captured from the camera. Used as a test program to verify if OpenCV has been properly installed.
-# ------------------------------------------------------------------------------
-# automaticdai
-# YF Robotics Labrotary
-# Instagram: yfrobotics
-# Twitter: @yfrobotics
-# Website: https://www.yfrl.org
-# ------------------------------------------------------------------------------
-
-import cv2
-import time
-import numpy as np
-
-CAMERA_DEVICE_ID = 0
-IMAGE_WIDTH = 320
-IMAGE_HEIGHT = 240
-fps = 0
-
-
-def visualize_fps(image, fps: int):
-    if len(np.shape(image)) < 3:
-        text_color = (255, 255, 255)  # white
-    else:
-        text_color = (0, 255, 0)  # green
-    row_size = 20  # pixels
-    left_margin = 24  # pixels
-
-    font_size = 1
-    font_thickness = 1
-
-    # Draw the FPS counter
-    fps_text = 'FPS = {:.1f}'.format(fps)
-    text_location = (left_margin, row_size)
-    cv2.putText(image, fps_text, text_location, cv2.FONT_HERSHEY_PLAIN,
-                font_size, text_color, font_thickness)
-
-    return image
-
-
-if __name__ == "__main__":
-    try:
-        # create video capture
-        cap = cv2.VideoCapture(CAMERA_DEVICE_ID)
-
-        # set resolution to 320x240 to reduce latency
-        cap.set(3, IMAGE_WIDTH)
-        cap.set(4, IMAGE_HEIGHT)
-
-        # Loop to continuously get images
-        while True:
-            # ----------------------------------------------------------------------
-            # record start time
-            start_time = time.time()
-
-            # Read the frames from a camera
-            _, frame = cap.read()
-
-            # show image
-            cv2.imshow('frame', visualize_fps(frame, fps))
-
-            # ----------------------------------------------------------------------
-            # record end time
-            end_time = time.time()
-
-            # calculate FPS
-            seconds = end_time - start_time
-            fps = 1.0 / seconds
-            print("Estimated fps:{0:0.1f}".format(fps))
-
-            # if key pressed is 'Esc' then exit the loop
-            if cv2.waitKey(33) == 27:
-                break
-    except Exception as e:
-        print(e)
-    finally:
-        # Clean up and exit the program
-        cv2.destroyAllWindows()
-        cap.release()
